==> src/18_psd_download_matrix.ipynb <==
# title: MapBiomas Soil
# subtitle: 18. Download PSD Design Matrix from GEE
# author: Alessandro Samuel-Rosa and Taciara Zborowski Horst
# data: 2025

# Load required libraries
import ee
import ee.batch

# Initialize the Earth Engine API
# ee.Authenticate()
ee.Initialize(project='mapbiomas-solos-workspace')

# Define the path to the FeatureCollection directory
fc_dir = 'projects/mapbiomas-workspace/SOLOS/AMOSTRAS/MATRIZES/collection3/'
fc_name = 'c03_pds_v2025_11_14b'
fc_path = f'{fc_dir}{fc_name}'
print(f'FeatureCollection path: {fc_path}')

# Define the output Google Drive folder
gdrive_folder = 'Earth Engine Exports'

# Export the FeatureCollection to Google Drive
fc = ee.FeatureCollection(fc_path)
task = ee.batch.Export.table.toDrive(
    collection=fc,
    description=f'ASR Export {fc_name} to Drive',
    folder=gdrive_folder,
    fileNamePrefix=fc_name,
    fileFormat='CSV'
)
task.start()
print(f"Export task started. Check Google Drive for file.")
print(task.status())
