==> src/18_psd_modeling.ipynb <==
# title: Download asset (data frame) from GEE
# author: Alessandro Samuel-Rosa, Taciara Zborowski Horst
# description: This script downloads a data frame from Google Earth Engine (GEE) and saves it as a
# CSV file in the "data" directory.
# date: 2025-04-25

# Load required libraries
import ee
import ee.batch

# Initialize the Earth Engine API
ee.Authenticate()
ee.Initialize(project='mapbiomas-solos-workspace')


# psd_c02beta_00_10cm_v2
fc = ee.FeatureCollection('projects/mapbiomas-workspace/SOLOS/AMOSTRAS/MATRIZES/granulometry/psd_c02beta_00_10cm_v2')
task = ee.batch.Export.table.toDrive(
    collection=fc,
    description='psd_c02beta_00_10cm_v2',
    folder='Earth Engine Exports',
    fileNamePrefix='psd_c02beta_00_10cm_v2',
    fileFormat='CSV'
)
task.start()
print(f"Export task started. Check Google Drive for file.")
print(task.status())
